Add --device parameter to pyhdk tests
